Add test for tooling invocations set up by CMake API
The purpose of the added test is to verify the way CMake sets up things
like import paths for tooling, in particular qmllint and qmlcachegen.
The test also ensures the application can actually run, which confirms
it can also load the QML module associated with the executable as well
as separate QML modules that have dependencies between them.

Task-number: QTBUG-95144
Pick-to: 6.2
Change-Id: Ice22dc4c9336f0271829e3abc21913355dcbd811
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>