Fix regression in ListView/Flickable event delivery
The regression was introduced with 2acb31641fc9c34d24ac29232cdfec

The problem was that when we entered
QQuickListViewPrivate::wantsPointerEvent(), the event was not localized
to the QQuickFlickable when mouse filtering was done.

Therefore, since the code assumed that the localPos was local to the
Flickable, it always ended up assuming that the point was within the
bounds of the header/footer.

Fixes: QTBUG-89409
Change-Id: Ib595e61b7995241b58e3051da09139e1e1f13cba
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit b2e4f09bfa2c42af562f6741c0a0246a53028f0a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>