Canvas: Add a means to override the DPR used via an environment variable
Overriding the DPR enables the system to not use as much memory and
resources that a low-end device or mobile might have a limited amount of
for rendering the Canvas item. In a situation where the user is aware of
the consequences and impact of forcing the DPR to be different to the
canvas window's DPR then this can be useful in those situations.

By having the environment variable to override the DPR then it makes it
possible for those needing this right away in all versions of Qt, so we
add this now with a view to add API for this at a later date.

Change-Id: Icafaa04a15678e2c7f24bc791026676f476a6ced
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit b0193908258f315a5736edcbc2a034d3194d42c9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>