Clean up arguments qt6_import_qml_plugins() passes to qmlimportscanner
We already provide the list of .qml files via the .qrc files, so we
don't need to explicitly list out all the .qml files on the command
line as well. Remove them.

The first argument given to qmlimportscanner is the source directory
of the target whose qml plugins are being imported. That argument
should be preceded by -rootPath, but it wasn't. It still worked due
to the way qmlimportscanner's argument parsing was implemented, but
we shouldn't rely on that. Precede it with -rootPath.

Fixes: QTBUG-95094
Change-Id: I863a1bda2e54bb70d2dabe43525ad972f8cdadfd
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit de2637d98d289f30cea6245baea27287ffc1675d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>