DefaultGroupBox: don't assign anything to a delegates x and y, but use insets
We should never assign anything to a delegate's x and y from
the style.  Those assignments will anyway be overwritten by
QQuickControl upon relayout. Instead, we need to use insets.

A regression from this is seen in the macOS style for 6.2, where
the GroupBox frame is positioned wrong (since the x and y bindings
set by the style are lost).

Pick-to: 6.2 6.1
Change-Id: Ifcc76af89fa2bc5ea9d03b175a2b56afa72fb204
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>