Make it possible to adopt an existing QRhi via QQuickGraphicsDevice
Just as QQuickRenderTarget allows with its (also \internal)
fromRhiRenderTarget to import (but not own) and existing
QRhiRenderTarget for a QQuickWindow, we can have the same for the QRhi
itself.

Having a fromRhi() in QQuickGraphicsDevice allows telling a
QQuickWindow to use an existing QRhi. This makes sense in particular
with QQuickRenderControl: one can now render two or more scenes (each
with its own QQuickWindow and QQuickRenderControl, but all using the
same QRhi under the hood) and then use the resulting textures with the
QRhi, because the QRhiTextures all belong to it to begin with.

While this is useful for applications in general, it is also an
enabler for modernizing QQuickWidget in the future, in case we choose
to do so.

Comes with an autotest case that, besides testing, is also a
demonstration of the feature. Once an application has rendered a frame
with the same QRhi from all the Qt Quick scenes, it can go on to do
something with the textures (e.g. compose them together in some other
window) because both belong to the same one QRhi.

Note that this comes with certain limitations, similarly to
QQuickGraphicsDevice::fromOpenGLContext. (the docs are noting this for
both functions now) For example, one cannot just pull out a QRhi from
under an on-screen QQuickWindow or QQuickView and expect it to be
usable in all cases, without limitations. For example, if the threaded
render loop is in use, using the QRhi extracted from the QQuickView is
a very bad idea.  However, in the common QQuickRenderControl-based
cases this is not an issue because there one would render frames from
the (offscreen) QQuickWindows one by one on the same thread.

Change-Id: I91b8b1f11f1cb4a84824c47d3b1d4f31833dd61b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>