QQuickBoundaryRule: Port from finalize callback to QQmlParserStatus
QQuickBoundaryRule only uses the finalize callback to set a flag when
the component is actually completed. This can be easily done with the
public QQmlParserStatus instead.
Besides being clearer thanks to using a documented approach, this should
also slightly improve performance as it avoids unnecessary work with the
meta-object system.

Task-number: QTBUG-96054
Change-Id: I2a4c8b3974f42aad5a47acb2c133aca035773873
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>