Use <major>.0 as default version for QML files
Using the full module versions presents problems if you bump the minor
version of your module. Bumping the minor version then bumps the
versions of all QML files not explicitly versioned. This is certainly
not what you want as the files should still be available to imports of
the old version.

Having it default to minor version 0 is more practical because many QML
files are available from version 0 of their respective modules. Now you
need to add version entries for files you add after .0 in order for them
not to show up in imports of earlier versions. This is less of a
problem, though, even if you forget it.

In addition, use PAST_MAJOR_VERSIONS to derive additional versions to be
added to the QML files. This allows us to drop a lot of boiler plate
code from our own modules.

Pick-to: 6.2
Change-Id: I8e4cfc16180af30e8bafc0a62137e9018f7eaee8
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>