QQuickAbstractAnimation: Avoid finalizer callback
We can observe the following: setRunning only installs a finalizer in
the case where the component has not completed yet. Therefore, it must
already be safe to execute the code in setRunning between
componentComplete and the time where the finalizers run. Thus, instead
of relying on finalizers, we can simply remember whether we have to do
some work, and do it in componentComplete.

Change-Id: I07f9c943517e8a9e8d8183743804c9f99488bd45
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>