Make TextField.implicitWidth fixed
Previously, implicitWidth was based on contentSize() (which was based
on the full textwidth inside the TextField)
This had the unfortunate consequence that an empty TextField (which is
very normal) was very narrow on startup (18 pixels on macOS style), but
would grow wider as the user kept on typing. When the TextField was in a
layout, this would in addition lead to a relayout, even though it
usually was barely noticeable. This change aligns that behavior to the
other non-native styles.

This also fixes a test on native style windows:
TextField::test_implicitWidth where it assumes that the implicitWidth of
TextField without a background is simply equal to left+right padding.
This failed because the calculation of the geometry of the placeholder
item was wrong.

Task-number: QTBUG-95679
Change-Id: Ie658411664d88df6cb6a337376bdcf897e8c1790
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit e0990f8b4069ca912de5a27df00a8039aa4a4731)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>