tst_combobox: skip indicator check for the native styles
The ComboBox::test_defaults() test checks that there is an
indictor present in the style. But for the native styles, the
indicator is drawn as a part of the background. So skip the
failing test when running on mac and windows.

Task-number: QTBUG-95678
Change-Id: I35f1900044bad1a519d9883b2061810bff8eb7a8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit 8ce979557c6048659d2baba587cf9369324b5a24)