Remove the per-rendercontext srb cache
Remove the srbCache which is effectively a map of unbounded size.
Depending on scene structure and changes to the scene that trigger
rebuilding the batching structures of the scenegraph, it can grow
unexpectedly large, potentially growing continuously on every frame,
with animated scenes that manage to trigger the "right" kind of
changes to the scene.

Other similar structures, such as the graphics pipeline or the sampler
tables have a known maximum size, because we know that there can only
be a certain number of combinations in the values that form the keys
to the map. Whereas with QRhiShaderResourceBindings objects the key is
the full binding list, including the QRhiBuffer/Texture/Sampler
pointers.

This becomes complex pretty quickly when combined with the internal
mechanisms of the scenegraph's batching system and the lifecycle of
these QRhiResource objects. For instance, uniform buffers live in
Batch, which is a pooled, (semi-)transient object: changes to the
scene that lead to full or partial rebuild will generate a new set of
Batches, meaning a QSGGeometryNode can be associated with a different
QRhiBuffer afterwards (not that we create new buffers all the time,
they are reused, but they can be associated with different nodes over
their lifetime). Which then means that to render the node, we now need
a different srb as well (that references the correct
QRhiBuffer). Depending on the scene and how it changes over time, this
may lead to an explosive number of combinations, thus a "continuously"
growing srbCache map, inserting more and more new entries, while
leaving perhaps-unused entries in there for ever. Trimming is far from
trivial as the costs of managing the purging of unused entries is
likely more expensive than not having the "cache" in the first place.

Thus, drop the whole data structure and give ownership of the srb to
Element (which is the object in the shadow tree for each
QSGGeometryNode). To not completely degrade performance, this needs
some new enablers in QRhi, namely the ability to have a fast path for
replacing the QRhiResources (buffers, textures, samplers) in the
binding list of an already baked srb without having to go through full
rebaking. (e.g. with Vulkan, we want to reuse all the existing
descriptor set layouts and the descriptor sets themselves, it's just
the descriptors themselves that need new values written)

So now, instead of looking up in a QHash, we first compare the full
binding list (this is what one gets with the QHash when there's a hit
anyway: at minimum the binding list is hashed and then compared). If
this matches then fine (and we saved the time spent on hashing even).
Then, if there's a mismatch, we now do a layout-only comparison. If
this matches, then we use the new enablers in QRhi to do a "light
update" (different resources, but same binding points, types, etc.) to
the srb. Finally, if there was a mismatch here too, we go through
create() to do a full rebuild.

Therefore, the performance effects of this patch depend heavily on the
scene. Scenes that do not do changes that trigger re-batching often
may see improvements even. Others are assumed to have small or
negligible consequences with regards to performance (due to now
spending time on a layout-compatible comparison of binding lists which
was not there before). In return, we do not need to have the srbCache
QHash at all, leading to, with scenes exhibiting the relevant
patterns, significantly reduced memory usage.

Once srbCache is gone, the following issues need to be handled:

First, now that the srb ownership stays with the shadow node
(Element), instead of guaranteed to stay around for ever in an srb
cache data structure, throwing references into the keys used to look
up pipelines is problematic (and was never nice to begin with).

Get rid of this, building on the new QRhi API that gives us a list of
uints that can be used to test for layout compatibility.  The
container is implicitly shared, which makes the switch quite simple
and efficient.

Second, all this does not mean that some sort of reuse of srb objects
is not needed. Consider scrolling a list view. Nodes (and so shadow
nodes, i.e. Element objects) come and go in this case. Creating full
new srb objects for the Elements for the list items becoming visible
is not ideal. (even if only really dubious with Vulkan, where there
are true Vulkan API objects underneath)

To facilitate reuse, reintroduce an srb pool in ShaderManager, but
this time with different semantics: this stores unused srb objects,
with the serialized layout description as the key. When a new Element
needs an srb, we try to pick a (layout compatible) srb from this
pool. If there is a match, the Element gets the srb with ownership,
and the srb is removed from the pool. When an Element is about to go
away, its srb may be "moved" to the pool when the pool is not too
large already.  The threshold for being too large is currently simply
1024 elements (in the QMultiHash). This can be overridden with an
environment variable, although that should pretty much be never needed
in practice.

Fixes: QTBUG-96130
Change-Id: Ie5a49beeb6dea0db6a81fdceebf39374ad192b0e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 84f30bf17170769795ddd8dbe1e7008ddbef2f6a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>