Shapes interactive example: use Controls canonically; architecture
As long as we had the funcs array, it would have been nice to be able to
directly get an index from whatever "tool" is selected. For example,
TabBar has currentIndex; but TabBar is otherwise not a great fit here.
But the funcs array wasn't so elegant anyway.

So that prompts switching to the known OO design pattern in which
drawing tools are objects containing functions that "do" the drawing, so
that there's no conditional dispatch: just ask the tool to handle each
mouse state change. I.e. here the DragHandler on the canvas uses the
tool's shapeType Component to create the shape, but the DragHandler
doesn't have to care which shape it is.

Handles for moving the shapes' control points are also shape-agnostic:
- a handle is a sort of template taking the path x and y properties to
  read and write
- the handle initializes its position from those properties
- the DragHandler inside gets to do what it does best: just drag the
  handle, nothing else (instead of needing to script the movement of
  both the handle and the control point, we don't script either one)
- declarative bindings on the handle's position update the path
  properties whenever the handle moves

And some styling:
- Switch looks better than Button for toggling
- use palette colors so that it looks good in dark mode too
- fix handle colors getting stuck on yellow
- borders on handles (yellow on white makes it hard to see the edge otherwise)
- reorder curve types by mathematical order (line, quadratic, cubic)

Change-Id: Iefd980f428601840deb55370aad9256748855f07
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit d5fb3c9a61d92bda7830ea3656fc44e83e7beae1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>