Doc: try to make "on" syntax more discoverable
Currently it's difficult for a user looking at the "on" syntax to find
anything online by searching for "on syntax qml". The information is
there, but it's not easily discoverable, as there's no way a user will
know to search for "Property Modifier Objects" or "Property Modifier
Types".

Change-Id: I453e8626ac5fe6656cd8e5adbcfc0f80faf4be9f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d3623d04ecceabff1a4a179079730102a804cae7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>