Make sure to install designer files with non-prefixed builds
It is not enough to just call qt_copy_or_install with a DESTINATION
value prefixed with only ${INSTALL_QMLDIR}/ because for non-prefixed
builds this will install relative to the directory this cmake file is
in. Instead we have to combine ${QT_INSTALL_DIR} with the target path to
make sure the final path is always correct. This pattern is used all
over, so we muse do it as well.

Change-Id: I571e19b42a611eaa1f368645e53114a99cb8fdaa
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit a5524a7544506e7915e9ab1d0ba3eaf687f401ed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>