Fix wrong assumption about TextField in tst_QQuickPopup::cursorShape()
The test assumed that an empty TextField was wider than 20 pixels.
This is not the case today on macOS native style.

TextField should probably also be made wider on macOS style, but
tst_QQuickPopup::cursorShape() should be more robust in order to
handle this. Afterall, it is not testing TextField, but simply using
it in order to test if the cursor shape changes if the hover is moved
away from the popup.

Task-number: QTBUG-95679
Pick-to: 6.2
Change-Id: I5130f161a5e92b032d228ffb515816c4043105af
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>