qmldom: improve newlines handling with AttributesSequence::Preserve
Improve the preservation of empty lines before an attribute
declaration when reformatting trying to preserve the original
attribute sequence.
Previously we inserted extra newlines when the type of object
attribute did change, in an attempt to create/preserve sections of
attributes with the same type.
The goal was not to change files following the coding guidelines too
much.
The problem was that it was invasive (might add a newline that was not
there), and actually incorrect for property declarations bound to a
value.
Tying to preserve the newlines in the input is actually a much better
policy.

Change-Id: I6bc95229cb7012243930b6b28aa457e331472de9
Pick-to: 6.2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>