Examples: Fix top level build
Examples using the "shared" pattern need to be recorded in the list in
examples/quick/CMakeLists.txt, so that we do the necessary AUTOMOC
handling for them in a top level build. For details, see the commit
which introduced that list.
Amends 34bc389543c37b8439c9c0b48f471649b6529868.

Pick-to: 6.2
Change-Id: I6945b0c8445fe436da259dd240342e1c3b5cbf26
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Craig Scott <craig.scott@qt.io>